requirements.txt
absl-py==2.1.0
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
executing==2.0.1
filelock==3.14.0
fonttools==4.51.0
fsspec==2024.3.1
grpcio==1.63.0
idna==3.7
ipython==8.24.0
jedi==0.19.1
Jinja2==3.1.4
kiwisolver==1.4.5
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mpmath==1.3.0
networkx==3.3
numpy==1.23.5
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
picamera==1.13
pillow==10.3.0
prompt-toolkit==3.0.43
protobuf==4.21.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
RPi.GPIO==0.7.1
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
smbus2==0.4.3
stack-data==0.6.3
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.3
